requirements.txt
flask
flask-restful
flask-jwt-extended
flask-sqlalchemy
uuid
flask-marshmallow
marshmallow
passlib
flask-uploads
python-dotenv
flask-migrate
httpie
requests
werkzeug==0.16.1
marshmallow-sqlalchemy
psycopg2
webargs
marshmallow-jsonapi
flask-caching
oauthlib==2.1.0
requests-oauthlib==1.1.0
flask-oauthlib
google-api-python-client
google-auth
flask-limiter
uwsgi